+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/providers/unstructured/images.py
@@ -1,34 +1,20 @@
+from collections.abc import Sequence
 from pathlib import Path
 
 from PIL import Image
-from pydantic import BaseModel
 from unstructured.chunking.basic import chunk_elements
 from unstructured.documents.elements import Element as UnstructuredElement
 from unstructured.documents.elements import ElementType
 
-from ragbits.core.llms.base import LLM, LLMType
-from ragbits.core.llms.factory import get_preferred_llm
-from ragbits.core.prompt import Prompt
+from ragbits.core.llms.base import LLM
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
-from ragbits.document_search.documents.element import Element, ImageElement
+from ragbits.document_search.documents.element import Element, IntermediateElement, IntermediateImageElement
 from ragbits.document_search.ingestion.providers.unstructured.default import UnstructuredDefaultProvider
 from ragbits.document_search.ingestion.providers.unstructured.utils import (
-    ImageDescriber,
     crop_and_convert_to_bytes,
     extract_image_coordinates,
     to_text_element,
 )
-
-DEFAULT_IMAGE_QUESTION_PROMPT = "Describe the content of the image."
-
-
-class _ImagePrompt(Prompt):
-    user_prompt: str = DEFAULT_IMAGE_QUESTION_PROMPT
-    image_input_fields: list[str] = ["images"]
-
-
-class _ImagePromptInput(BaseModel):
-    images: list[bytes]
 
 
 class UnstructuredImageProvider(UnstructuredDefaultProvider):
@@ -64,17 +50,18 @@ class UnstructuredImageProvider(UnstructuredDefaultProvider):
             llm: llm to use
         """
         super().__init__(partition_kwargs, chunking_kwargs, api_key, api_server, use_api)
-        self.image_describer: ImageDescriber | None = None
         self._llm = llm
 
     async def _chunk_and_convert(
         self, elements: list[UnstructuredElement], document_meta: DocumentMeta, document_path: Path
-    ) -> list[Element]:
+    ) -> Sequence[Element | IntermediateElement]:
         image_elements = [e for e in elements if e.category == ElementType.IMAGE]
         other_elements = [e for e in elements if e.category != ElementType.IMAGE]
         chunked_other_elements = chunk_elements(other_elements, **self.chunking_kwargs)
 
-        text_elements: list[Element] = [to_text_element(element, document_meta) for element in chunked_other_elements]
+        text_elements: list[Element | IntermediateImageElement] = [
+            to_text_element(element, document_meta) for element in chunked_other_elements
+        ]
         if self.ignore_images:
             return text_elements
         return text_elements + [
@@ -83,7 +70,7 @@ class UnstructuredImageProvider(UnstructuredDefaultProvider):
 
     async def _to_image_element(
         self, element: UnstructuredElement, document_meta: DocumentMeta, document_path: Path
-    ) -> ImageElement:
+    ) -> IntermediateImageElement:
         top_x, top_y, bottom_x, bottom_y = extract_image_coordinates(element)
         image = self._load_document_as_image(document_path)
         top_x, top_y, bottom_x, bottom_y = self._convert_coordinates(
@@ -91,13 +78,7 @@ class UnstructuredImageProvider(UnstructuredDefaultProvider):
         )
 
         img_bytes = crop_and_convert_to_bytes(image, top_x, top_y, bottom_x, bottom_y)
-        prompt = _ImagePrompt(_ImagePromptInput(images=[img_bytes]))
-        if self.image_describer is None:
-            llm_to_use = self._llm if self._llm is not None else get_preferred_llm(LLMType.VISION)
-            self.image_describer = ImageDescriber(llm_to_use)
-        image_description = await self.image_describer.get_image_description(prompt=prompt)
-        return ImageElement(
-            description=image_description,
+        return IntermediateImageElement(
             ocr_extracted_text=element.text,
             image_bytes=img_bytes,
             document_meta=document_meta,