+++ packages/ragbits-core/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add support for images in few shot prompts (#155)
 - BREAKING CHANGE: Vector Stores are now responsible for creating embeddings (#288)
 - Add instruction on how to use local servers for LLMs (#395).
+- Introduce intermediate image elements (#139)
 
 ## 0.9.0 (2025-02-25)