+++ packages/ragbits-document-search/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Compability with the new Vector Store interface from ragbits-core (#288)
 - Fix docstring formatting to resolve Griffe warnings
+- Introduce intermediate image elements (#139)
 
  - added traceable